requirements.txt
msrest>=0.6.21
pytz>=2021.1
pillow~=11.1.0
azure-core~=1.32.0